% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GForecast.R
\name{GForecast}
\alias{GForecast}
\title{Multidirectional Group Time-Series Forecasting'}
\usage{
GForecast(data, target, groupId, date, type = "ets", h = 12,
  fitted = FALSE)
}
\arguments{
\item{data}{"A data frame that includes time, group, and target variable to forecast."}

\item{target}{"Numeric vector containing the values to forecast."}

\item{groupId}{"Grouping variable in the data frame."}

\item{date}{"Name of the date column. Must be in '%Y-%m-%d' format."}

\item{type}{"'ets' or 'auto.arima'. Default = 'ets'.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item groupId
 \item date
 \item key
 \item target
 \item lo.80
 \item lo.95
 \item hi.80
 \item hi.95
}
}
\description{
Description: Apply ARIMA or ETS to grouped time series data and specify direction of horizon.
}
\details{
No further details.
}
\examples{
GForecast(home_sales, group = "zip", target = "medianSalePrice")
}
\author{
James Craig
}
